update: verified fixes for 11 concurrency/resilience bugs

+++ src/main/java/com/boombapcompile/blckvox/service/audio/AudioDurationCalculator.java
@@ -27,7 +27,7 @@ public final class AudioDurationCalculator {
         if (pcmByteLength <= 0) {
             return 0.0;
         }
-        return (double) pcmByteLength / (sampleRate * bytesPerSample * channels) * 1000.0;
+        return (double) pcmByteLength / ((long) sampleRate * bytesPerSample * channels) * 1000.0;
     }
 
     /**
